+++ src/modules/artist/dto/createArtist.dto.ts
